refactor: FarePolicy builder 생성자로 수정

+++ src/main/java/wooteco/subway/fare/domain/farestrategy/LineFare.java
@@ -13,6 +13,6 @@ public class LineFare implements FareStrategy {
 
     @Override
     public Money calculate(Money value) {
-        return value.add(BigDecimal.valueOf(lineExtraFare));
+        return value.add(lineExtraFare);
     }
 }
